Added recursive evaluation of expressions
Some expressions may contain references, e.g. '17628this', or operators, e.g. '',
in their expression values even if the expression key does not start with a '$'.
This case was previously not covered by 'evaluateDocumentExpression' as it used
the literal expression value if the expression key did not start with a '$'. The
test which was added to 'testEvaluateReduce' verifies that the implementation works
and thus evaluates nested expressions. The example for the test is from the
official MongoDB documentation for ''. See https://docs.mongodb.com/manual/reference/operator/aggregation/reduce/.